requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
crewai>=0.148.0
anthropic>=0.25.0
openai>=1.13.3
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
pandas>=2.1.0
schedule>=1.2.0
aiofiles>=23.2.1
httpx>=0.25.2
Pillow>=10.1.0
sqlalchemy>=2.0.23
alembic>=1.13.0